agregar y editar tanto empleados como productos, cambio BD ahora precio es float, DELETE NO FUNCIONA mejor hay que desactivar el usuario o agotar el producto

+++ server/init_db/init_db.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS `modulasys`.`inventarios` (
   `nombreProducto` VARCHAR(45) NOT NULL,
   `existenciaProducto` INT NOT NULL,
   `urlFotoProducto` VARCHAR(45) NOT NULL,
-  `precioProducto` INT NOT NULL,
+  `precioProducto` DECIMAL(10,2) NOT NULL,
   PRIMARY KEY (`idinventarios`))
 ENGINE = InnoDB;
